Fixed folder and file names to match; added README
This is really about adding a README that allows this to quickly be referenced so they can be edited.  The entire goal of this project is to make it so these different GitHub Action files can be made consistent, by having a quick way to link to each you can then modify them to be consistent.